Support multiple statemachines
