+++ addons/civilian_statemachine/XEH_postInit.sqf
@@ -1,5 +1,5 @@
 #include "script_component.hpp"
 
 if (isServer) then {
-    [configFile >> "CfgSerialKillers" >> "Civilian_Statemachine"] call CBA_statemachine_fnc_createFromConfig;
+    [configFile >> "CfgSerialKillers" >> "CivilianMovement"] call CBA_statemachine_fnc_createFromConfig;
 };
